Move hashivault_read definition to module_utils
